LayoutTests/fast/forms/time/time-editable-components/time-editable-components-second-and-millisecond-field-expected.txt

Test for presence of second and millisecond fields in <input type=time>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Default
PASS numberOfFields is 3

Zero seconds
PASS numberOfFields is 3

Zero seconds and zero milliseconds
PASS numberOfFields is 3

Zero seconds and non-zero milliseconds
PASS numberOfFields is 5

Non-zero seconds
PASS numberOfFields is 4

Non-zero seconds and zero milliseconds
PASS numberOfFields is 4

Non-zero seconds and non-zero milliseconds
PASS numberOfFields is 5

Step attribute minute precision
PASS numberOfFields is 3

Step attribute second precision
PASS numberOfFields is 4

Step attribute millisecond precision
PASS numberOfFields is 5

PASS successfullyParsed is true

TEST COMPLETE
